refactor lora related parts
