+++ verl_omni/workers/engine/fsdp/diffusers_impl.py
@@ -24,7 +24,6 @@ from typing import Callable, Optional
 
 import torch
 import torch.distributed
-from peft import LoraConfig
 from tensordict import TensorDict
 from torch.distributed.fsdp import FullyShardedDataParallel as FSDP
 from torch.distributed.fsdp.api import FullStateDictConfig, ShardedStateDictConfig, StateDictType
@@ -53,7 +52,7 @@ from verl.utils.fsdp_utils import (
 )
 from verl.utils.memory_utils import aggressive_empty_cache
 from verl.utils.model import convert_weight_keys
-from verl.utils.py_functional import append_to_dict, convert_to_regular_types
+from verl.utils.py_functional import append_to_dict
 from verl.workers.config import FSDPEngineConfig, FSDPOptimizerConfig
 from verl.workers.engine.base import BaseEngine, BaseEngineCtx, EngineRegistry
 from verl.workers.engine.fsdp.utils import create_device_mesh, get_sharding_strategy
@@ -68,6 +67,7 @@ from verl_omni.pipelines.utils import (
 )
 from verl_omni.utils.fsdp_utils import collect_lora_params
 from verl_omni.workers.config import DiffusionModelConfig
+from verl_omni.workers.engine.lora_adapter_mixin import LoRAAdapterMixin
 
 logger = logging.getLogger(__file__)
 logger.setLevel(os.getenv("VERL_LOGGING_LEVEL", "WARN"))
@@ -75,7 +75,7 @@ logger.setLevel(os.getenv("VERL_LOGGING_LEVEL", "WARN"))
 device_name = get_device_name()
 
 
-class DiffusersFSDPEngine(BaseEngine, ABC):
+class DiffusersFSDPEngine(LoRAAdapterMixin, BaseEngine, ABC):
     """Base Diffusers engine using PyTorch FullyShardedDataParallel (FSDP).
 
     Supports model sharding, activation/optimizer offloading, LoRA, and sequence parallelism.
@@ -228,42 +228,6 @@ class DiffusersFSDPEngine(BaseEngine, ABC):
 
             module.can_generate = lambda: False
             module.config.save_pretrained = save_config.__get__(module.config)
-
-        return module
-
-    def _build_lora_module(self, module):
-        lora_adapter_path = getattr(self.model_config, "lora_adapter_path", None)
-        policy_state_adapters = tuple(getattr(self.model_config, "policy_state_adapters", ("default",)))
-        extra_adapters = tuple(adapter for adapter in policy_state_adapters if adapter not in ("default", "reference"))
-        if lora_adapter_path is not None:
-            from verl.utils.fs import copy_to_local
-
-            print(f"Loading pre-trained LoRA adapter to from: {lora_adapter_path}")
-            # Copy adapter to local if needed
-            local_adapter_path = copy_to_local(lora_adapter_path, use_shm=self.model_config.use_shm)
-
-            module.load_lora_adapter(local_adapter_path)
-            peft_config = getattr(module, "peft_config", {}).get("default", None)
-            for adapter_name in extra_adapters:
-                if peft_config is not None and adapter_name not in getattr(module, "peft_config", {}):
-                    module.add_adapter(peft_config, adapter_name=adapter_name)
-        else:
-            # Convert config to regular Python types before creating PEFT model
-            lora_config = {
-                "r": self.model_config.lora_rank,
-                "lora_alpha": self.model_config.lora_alpha,
-                "init_lora_weights": self.model_config.lora_init_weights,
-                "target_modules": convert_to_regular_types(self.model_config.target_modules),
-                "target_parameters": convert_to_regular_types(self.model_config.target_parameters),
-                "exclude_modules": convert_to_regular_types(self.model_config.exclude_modules),
-                "bias": "none",
-            }
-            module.add_adapter(LoraConfig(**lora_config), adapter_name="default")
-            for adapter_name in extra_adapters:
-                module.add_adapter(LoraConfig(**lora_config), adapter_name=adapter_name)
-
-        if "default" in policy_state_adapters and hasattr(module, "set_adapter"):
-            module.set_adapter("default")
 
         return module
 
@@ -626,85 +590,21 @@ class DiffusersFSDPEngine(BaseEngine, ABC):
         if self._is_offload_optimizer:
             offload_fsdp_optimizer(self.optimizer)
 
-    def _set_adapter(self, name: str):
-        module = getattr(self.module, "_fsdp_wrapped_module", self.module)
-        if hasattr(module, "set_adapter"):
-            module.set_adapter(name)
-        elif hasattr(self.module, "set_adapter"):
-            self.module.set_adapter(name)
-        else:
-            raise AttributeError(f"Module does not support set_adapter({name!r})")
+    def _unwrap_adapter_module(self, module):
+        return getattr(module, "_fsdp_wrapped_module", module)
 
     @contextmanager
-    def use_adapter(self, name: str):
-        """Temporarily select a named PEFT adapter."""
-        self._set_adapter(name)
-        try:
-            yield
-        finally:
-            self._set_adapter("default")
-
-    def _active_adapter_trainable_params(self, adapter_name: str) -> list[torch.nn.Parameter]:
-        peft_model = getattr(self.module, "_fsdp_wrapped_module", self.module)
-        if not hasattr(peft_model, "set_adapter"):
-            raise AttributeError("Module does not support PEFT adapter selection.")
-        peft_model.set_adapter(adapter_name)
-        return list(filter(lambda param: param.requires_grad, peft_model.parameters()))
-
-    def copy_adapter(self, source: str = "default", target: str = "old") -> None:
-        """Copy LoRA state between named policy adapters."""
+    def _adapter_state_context(self):
         is_fsdp_module = fsdp_version(self.module) == 1
         is_offload_param = getattr(self, "_is_offload_param", False)
         origin_module_device = next(self.module.parameters()).device.type
         if is_fsdp_module and (is_offload_param or origin_module_device == "cpu"):
             load_fsdp_model_to_gpu(self.module)
-        summon_ctx = (
-            FSDP.summon_full_params(self.module, writeback=True, recurse=True) if is_fsdp_module else nullcontext()
-        )
 
+        ctx = FSDP.summon_full_params(self.module, writeback=True, recurse=True) if is_fsdp_module else nullcontext()
         try:
-            with summon_ctx, torch.no_grad():
-                source_params = self._active_adapter_trainable_params(source)
-                target_params = self._active_adapter_trainable_params(target)
-                if len(source_params) != len(target_params) or not source_params:
-                    raise ValueError(
-                        f"Adapter copy {source!r} -> {target!r} found mismatched params: "
-                        f"{len(source_params)} vs {len(target_params)}"
-                    )
-                for source_param, target_param in zip(source_params, target_params, strict=True):
-                    target_param.data.copy_(source_param.detach().data)
-        finally:
-            self._set_adapter("default")
-            if is_offload_param:
-                offload_fsdp_model_to_cpu(self.module)
-                aggressive_empty_cache(force_sync=True)
-
-    def ema_update_adapter(self, source: str = "default", target: str = "old", decay: float = 0.0) -> None:
-        """EMA-update target adapter parameters from source adapter parameters."""
-        if not 0.0 <= decay <= 1.0:
-            raise ValueError(f"Adapter EMA decay must be in [0, 1], got {decay}.")
-        is_fsdp_module = fsdp_version(self.module) == 1
-        is_offload_param = getattr(self, "_is_offload_param", False)
-        origin_module_device = next(self.module.parameters()).device.type
-        if is_fsdp_module and (is_offload_param or origin_module_device == "cpu"):
-            load_fsdp_model_to_gpu(self.module)
-        summon_ctx = (
-            FSDP.summon_full_params(self.module, writeback=True, recurse=True) if is_fsdp_module else nullcontext()
-        )
-
-        try:
-            with summon_ctx, torch.no_grad():
-                source_params = self._active_adapter_trainable_params(source)
-                target_params = self._active_adapter_trainable_params(target)
-                if len(source_params) != len(target_params) or not source_params:
-                    raise ValueError(
-                        f"Adapter EMA {source!r} -> {target!r} found mismatched params: "
-                        f"{len(source_params)} vs {len(target_params)}"
-                    )
-                for source_param, target_param in zip(source_params, target_params, strict=True):
-                    target_param.data.copy_(
-                        target_param.detach().data * decay + source_param.detach().clone().data * (1.0 - decay)
-                    )
+            with ctx:
+                yield
         finally:
             self._set_adapter("default")
             if is_offload_param:
@@ -722,7 +622,7 @@ class DiffusersFSDPEngine(BaseEngine, ABC):
 
         peft_config = None
 
-        peft_model = getattr(self.module, "_fsdp_wrapped_module", self.module)
+        peft_model = self._unwrap_adapter_module(self.module)
         if hasattr(peft_model, "peft_config"):  # LoRA
             peft_config = peft_model.peft_config.get("default", None)
             adapter_ctx = self.use_adapter(adapter_name) if adapter_name is not None else nullcontext()
@@ -739,7 +639,7 @@ class DiffusersFSDPEngine(BaseEngine, ABC):
         else:
             params = self.module.state_dict()
 
-        params = convert_weight_keys(params, getattr(self.module, "_fsdp_wrapped_module", self.module))
+        params = convert_weight_keys(params, self._unwrap_adapter_module(self.module))
 
         log_gpu_memory_usage("Before offload_fsdp_model_to_cpu", logger=logger)
         if self._is_offload_param:
@@ -786,14 +686,6 @@ class DiffusersFSDPEngine(BaseEngine, ABC):
             embeds = torch.nn.functional.pad(embeds, (0, 0, 0, pad_len))
             mask = torch.nn.functional.pad(mask, (0, pad_len))
         return embeds, mask
-
-    @contextmanager
-    def disable_adapter(self):
-        try:
-            self.module.disable_adapters()
-            yield
-        finally:
-            self.module.enable_adapters()
 
     def postprocess_batch_func(self, output_lst, indices, data: TensorDict):
         model_output = {}
